Fix: Typescript config
Add: Database connection
