+++ .gitignore
@@ -2,4 +2,4 @@ node_modules/
 .env/
 .idea/
 TP.md
-dist/
+build/
