--- middlewares/docs.middleware.ts
@@ -1,1 +1,0 @@
-export
