+++ package.json
@@ -7,14 +7,15 @@
     "test": "jest --verbose",
     "test-watch": "jest --verbose --watchAll",
     "dev": "concurrently -k -n \"Typescript,Node\" -p \"[{name}]\" -c \"blue,green\" \"tsc --watch\" \"nodemon dist/src/index.js\"",
-    "start": "tsc && node dist/src/index.js"
+    "start": "tsc-watch --onSuccess \"node ./build/index.js\""
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/mongoose": "^5.11.97",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "mongodb": "^5.6.0",
+    "mongoose": "^7.4.1",
     "nodemon": "^2.0.22",
     "swagger-jsdoc": "^5.0.1",
     "swagger-ui-express": "^4.6.3"
@@ -28,6 +29,7 @@
     "concurrently": "^8.2.0",
     "jest": "^29.5.0",
     "ts-jest": "^29.1.0",
+    "tsc-watch": "^6.0.4",
     "typescript": "^5.1.3"
   }
 }
